+++ src/index.ts
@@ -25,6 +25,7 @@ import { evidenceCommand } from "./commands/evidence.js";
 import { docsCommand } from "./commands/docs.js";
 import { analyticsCommand } from "./commands/analytics.js";
 import { flowCommand } from "./commands/flow.js";
+import { factoryCommand } from "./commands/factory.js";
 import { VERSION } from "./lib/constants.js";
 
 const program = new Command();
@@ -58,5 +59,6 @@ program.addCommand(evidenceCommand);
 program.addCommand(docsCommand);
 program.addCommand(analyticsCommand);
 program.addCommand(flowCommand);
+program.addCommand(factoryCommand);
 
 program.parse(process.argv);